+++ actix-settings/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Use new feature named `tls` for TLS settings. [#380]
+- Add new feature named `openssl` for TLS settings using `OpenSSL` dependency. [#380]
 - Add new function `settings::tls::Tls::get_ssl_acceptor_builder()` to build `openssl::ssl::SslAcceptorBuilder`. [#380]
 - Implement TLS logic for `ApplySettings<S>::try_apply_settings()`. [#380]
 - Add `openssl` dependency;